Finally fixed #72888, #73404, #75558: much more robust conversion of old config files, now hopefully absolutely no black notes anymore.
The reason for those was partly #68127 - a note's config was deleted, the
note itself was kept. The next time KNotes is used it load the journal but
obviously can't find the config. On quit only the minimal parameters were
written (no version information) and the following restart of KNotes it
thinks it found a KDE 1.x configuration. This is now fixed by examining
the config files a bit better.

CCMAIL: 72888-done@bugs.kde.org, 73404-done@bugs.kde.org, 75558-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=294818