Add DCOP accessors and setters for the note colors.
The signal sigDataChanged() has to be emitted instead of calling saveData()
directly to prevent the CTOR from saving the yet empty notes back to disk---at
that time the signal is not connected yet; additionally, saveData() would call
Journal::setDescription() with an empty editor, effectively deleting the note
before it was even read at all.

Journal::updated() has to be called to tell the resource that the data has
changed. setCustomProperty() does not do that, thus the color changes wouldn't
be saved to disk if nothing else was changed.

CCMAIL: caio1982@gmail.com


svn path=/branches/KDE/3.5/kdepim/; revision=550490